CPT Unit Test and moving TVMaze Proper
