+++ tests/components/test-cpts.php
@@ -9,8 +9,11 @@
 
 namespace LWTV\Tests;
 
-use LWTV\_Components\CPTs;
 use LWTV\Plugin;
+use LWTV\CPTs\Actors;
+use LWTV\CPTs\Characters;
+use LWTV\CPTs\Shows;
+use LWTV\CPTs\TVMaze;
 
 /**
  * CPTs Tests.
@@ -21,10 +24,10 @@ class CPTs_Test extends \WP_UnitTestCase {
 	 * Test if the post types exists
 	 */
 	public function test_cpts_exists() {
-		$actors     = post_type_exists( 'post_type_actors' );
-		$characters = post_type_exists( 'post_type_characters' );
-		$shows      = post_type_exists( 'post_type_shows' );
-		$tvmaze     = post_type_exists( 'post_type_tvmaze' );
+		$actors     = post_type_exists( Actors::SLUG );
+		$characters = post_type_exists( Characters::SLUG );
+		$shows      = post_type_exists( Shows::SLUG );
+		$tvmaze     = post_type_exists( TVMaze::SLUG );
 		$fake       = post_type_exists( 'post_type_faker' );
 
 		$this->assertFalse( $fake );
@@ -33,4 +36,47 @@ class CPTs_Test extends \WP_UnitTestCase {
 		$this->assertTrue( $shows );
 		$this->assertTrue( $tvmaze );
 	}
+
+	/**
+	 * Test if the ACTOR taxonomies exist
+	 */
+	public function test_actor_taxonomies_exists() {
+
+		foreach( Actors::ALL_TAXONOMIES as $taxonomy => $items ) {
+			$taxonomy_exists = taxonomy_exists( $taxonomy );
+			$this->assertTrue( $taxonomy_exists );
+		}
+
+		$fake_taxonomy = taxonomy_exists( 'lezactors_fake' );
+		$this->assertFalse( $fake_taxonomy );
+	}
+
+	/**
+	 * Test if the Character taxonomies exist
+	 */
+	public function test_character_taxonomies_exists() {
+
+		foreach( Characters::ALL_TAXONOMIES as $taxonomy => $items ) {
+			$taxonomy_exists = taxonomy_exists( $taxonomy );
+			$this->assertTrue( $taxonomy_exists );
+		}
+
+		$fake_taxonomy = taxonomy_exists( 'lezactors_fake' );
+		$this->assertFalse( $fake_taxonomy );
+	}
+
+	/**
+	 * Test if the Show taxonomies exist
+	 */
+	public function test_show_taxonomies_exists() {
+
+		foreach( Shows::ALL_TAXONOMIES as $taxonomy => $items ) {
+			$taxonomy_exists = taxonomy_exists( $taxonomy );
+			$this->assertTrue( $taxonomy_exists );
+		}
+
+		$fake_taxonomy = taxonomy_exists( 'lezactors_fake' );
+		$this->assertFalse( $fake_taxonomy );
+	}
+
 }
